[ch219] Change file to .ts and refine codes
