+++ src/pages/api/products/index.tsx
@@ -4,6 +4,7 @@ import { product } from '../../../entity/entities'
 const productHandler = (req: NextApiRequest, res: NextApiResponse) => {
   const { method, body } = req
 
+  // TODO: Handle the function properly
   const handlePatch = () => {
     return {
       message: 'This is a PATCH request, currently in progress...',
@@ -11,8 +12,9 @@ const productHandler = (req: NextApiRequest, res: NextApiResponse) => {
     }
   }
 
+  // TODO: Handle the function properly
   const handlePost = () => {
-    const products = [product('shoe1', 1), product('shoe2', 1), product('shoe3', 1)]
+    const products: Array<ReturnType<typeof product>> = [product('shoe1', 1), product('shoe2', 1), product('shoe3', 1)]
 
     return {
       products,
