+++ src/pages/api/users/index.tsx
@@ -5,6 +5,7 @@ import { user, product } from '../../../entity/entities'
 const userHandler = (req: NextApiRequest, res: NextApiResponse) => {
   const { method, body } = req
 
+  // TODO: Handle the function properly
   const handlePatch = () => {
     return {
       message: 'This is a PATCH request, currently in progress...',
@@ -12,6 +13,7 @@ const userHandler = (req: NextApiRequest, res: NextApiResponse) => {
     }
   }
 
+  // TODO: Handle the function properly
   const handlePost = () => {
     return {
       users: mockUsers(),
@@ -20,12 +22,12 @@ const userHandler = (req: NextApiRequest, res: NextApiResponse) => {
   }
 
   const mockUsers = () => {
-    const userNames = ['mocking-bird', 'wood-pecker', 'black-crow']
-    const products = [product('shoe1', 1), product('shoe2', 1), product('shoe3', 1)]
-    const histories = products.reverse()
-    const users = userNames.map(name => {
-      const password = [...name].reverse().join('')
-      const splitName = name.split('-')
+    const userNames: string[] = ['mocking-bird', 'wood-pecker', 'black-crow']
+    const products: Array<ReturnType<typeof product>> = [product('shoe1', 1), product('shoe2', 1), product('shoe3', 1)]
+    const histories: Array<ReturnType<typeof product>> = products.reverse()
+    const users: Array<ReturnType<typeof user>> = userNames.map(name => {
+      const password: string = [...name].reverse().join('')
+      const splitName: string[] = name.split('-')
 
       return user(name, password, splitName[0], splitName[1], '0', Role.User, products, histories)
     })
